сделала страницу home адаптивной
