+++ src/components/UI/NavBar/index.ts
@@ -1,0 +1,1 @@
+export { NavBar } from './NavBar'
