+++ src/components/catalog/Catalog.module.css
@@ -11,11 +11,41 @@
 }
 
 .catalog__products {
-    display: flex;
-    flex-wrap: wrap;
-    align-content: start;
+
+    display:grid;
+    grid-template-columns: repeat(3, minmax(300px, 1fr));
     margin-bottom: 40px;
     gap: 20px;
+}
+
+@media (max-width:1370px){
+    .catalog__products{
+        grid-template-columns: repeat(3, minmax(300px, 1fr));
+    }
+}
+
+@media (max-width: 1220px) {
+    .catalog__products {
+        grid-template-columns: repeat(2, minmax(300px, 1fr));
+    }
+}
+
+@media (max-width: 800px) {
+    .catalog__products {
+        grid-template-columns: repeat(1, minmax(300px, 1fr));
+    }
+}
+
+@media (max-width: 350px) {
+    .catalog__products {
+        grid-template-columns: repeat(1, minmax(280px, 1fr));
+    }
+}
+
+@media (max-width: 320px) {
+    .catalog__products {
+        grid-template-columns: repeat(1, minmax(268px, 1fr));
+    }
 }
 
 .catalog__button {
