+++ src/components/pages/Home/Home.tsx
@@ -1,5 +1,5 @@
 import {Header} from '../../header'
-import { NavBar } from "../../NavBar/NavBar";
+import { NavBar } from "../../UI/NavBar/NavBar";
 import { Catalog } from "../../catalog/Catalog";
 
 export const Home = () => {
