+++ src/style/container.css
@@ -10,7 +10,7 @@
 }
 
 .container_header {
-    margin: 0 auto;
+    margin: 0;
     width: 100%;
     height: 500px;
     border: none;
@@ -28,4 +28,28 @@
     width: 100%;
     /* max-height: 650px; */
     color: white;
+}
+
+@media (max-width:1200px){
+    .container{
+        max-width:1010px;
+    }
+}
+
+@media (max-width:1000px){
+    .container{
+        max-width:810px;
+    }
+}
+
+@media (max-width:800px){
+    .container{
+        max-width:610px;
+    }
+}
+
+@media (max-width:600px){
+    .container{
+        max-width:410px;
+    }
 }
